refactor(simulation): split simulation response mappers by feature
