+++ src/main/java/com/renewsim/backend/simulation_service/dashboard/web/SimulationDashboardController.java
@@ -4,7 +4,6 @@ import com.renewsim.backend.simulation_service.application.dashboard.GetPortfoli
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContext;
 import com.renewsim.backend.simulation_service.shared.web.SimulationRequestContextFactory;
 import com.renewsim.backend.simulation_service.dashboard.web.dto.PortfolioDashboardResponseDTO;
-import com.renewsim.backend.simulation_service.web.controller.SimulationWebMapper;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -20,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class SimulationDashboardController {
 
     private final GetPortfolioDashboardUseCase getPortfolioDashboardUseCase;
-    private final SimulationWebMapper webMapper = new SimulationWebMapper();
+    private final SimulationDashboardWebMapper dashboardWebMapper = new SimulationDashboardWebMapper();
     private final SimulationRequestContextFactory requestContextFactory = new SimulationRequestContextFactory();
 
     @Operation(summary = "Get executive portfolio dashboard for authenticated user")
@@ -30,6 +29,6 @@ public class SimulationDashboardController {
         SimulationRequestContext requestContext = requestContextFactory.from(auth);
 
         return ResponseEntity
-                .ok(webMapper.toWebDashboard(getPortfolioDashboardUseCase.getDashboard(requestContext.username())));
+                .ok(dashboardWebMapper.toWebDashboard(getPortfolioDashboardUseCase.getDashboard(requestContext.username())));
     }
 }
